feat: Add containerization support (#29)
Merge PR #28 

* feat: Add containerization support

Adds support for containerizing the app using docker

* fix(docker): Remove EXPOSE instruction to allow port configuration in Compose

Removing the EXPOSE instruction from the Dockerfile allows the
port mapping to be fully managed within the Docker Compose file.
This provides greater flexibility in configuring the port mapping
for different environments.

* fix(docker): Improve development experience with watch and Compose port config

Added a watch configuration to improve the development workflow and
removed environment variables related to the application port to
allow port configuration exclusively via Docker Compose. This
provides greater flexibility for different development environments.

* fix(swagger): Correct Swagger path for API documentation

Corrects the Swagger path configuration to ensure the API
documentation is accessible. The previous path was incorrect,
preventing the documentation from displaying.

* fix(config): Relax validation for default port and app URL configuration

Relaxes the validation rules for the default port and app URL
configuration to prevent warnings during startup when explicit
configurations are not provided. These defaults are only used in
the absence of explicit configurations.

* ci(docker): Add Docker image build verification to pull requests

Adds a CI job to verify that the Docker image builds successfully
on each pull request. This ensures that changes do not introduce
build errors in the Docker image and provides early feedback to
developers.

* fix(docker): Add restart policy to container configuration

Adds a restart policy to the container configuration to ensure
the container restarts automatically in case of failures. This
improves the resilience and availability of the application.

* fix(docker): Suppress warnings during Docker image build

Resolves warnings that were occurring during the Docker image build
process. These warnings did not prevent the image from building, but
have been addressed to improve build cleanliness and maintainability

* fix(docker): Allow external port configuration via environment variable

Allows configuring the external port for the Docker container using
an environment variable. This provides greater flexibility for
deployments and simplifies configuration in different environments.

* fix(docker): Add custom container names for improved readability

Adds custom names to Docker containers to improve readability and
facilitate easier identification and management. This makes it
easier to distinguish between different containers and simplifies
debugging and monitoring.

* chore(docs): Update README with Docker support instructions

Updates the README file to include instructions and a guide on how
to use Docker with the project. This provides developers with
clear guidance on setting up and running the application using
Docker.

* chore: update Developer tools information

* ci: Exclude README changes from triggering CI pipeline

Excludes changes to README files from triggering the CI pipeline.
Since README changes typically do not require testing or building,
this prevents unnecessary CI runs and saves resources.

* fix(ci): resolve warning about incorrect use of node environment

---

closes #28, #29